requirements.txt
python-telegram-bot==20.7
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0